[errors] moved errors to Errors.ts
