+++ test/EthWallet/EthUtils.spec.ts
@@ -1,4 +1,4 @@
-import {EthKeyAlreadyExist, EthKeyDoesNotExist, EthUtils} from "../../src/EthWallet/EthUtils";
+import {EthUtils} from "../../src/EthWallet/EthUtils";
 import {SecureStorageInterface} from "../../src/SecureStorage/SecureStorageInterface";
 import {NodeJsSecureStorage} from '../../src/SecureStorage/NodeJsSecureStorage'
 import {instance, mock, when} from "ts-mockito";
@@ -7,6 +7,7 @@ import {AES, enc} from 'crypto-js'
 import {mnemonicToEntropy, entropyToMnemonic, validateMnemonic} from 'bip39'
 import {unlinkSync} from "fs";
 import {isValidPrivate, isValidAddress} from "ethereumjs-util";
+import {EthKeyAlreadyExist, EthKeyDoesNotExist} from "../../src/Errors";
 
 const PRIVATE_TEST_KEY = "9b4b3da7d2d1c8749743a4ac17c151405f7900832bb4cf88735721cef4627096";
